Merging seaice_growth codes -- preliminary steps.
Done with this part!

This part, as intended, has not changed anything to the actual computation
results (lab_sea ad&fwd). It was "simply" a matter untying the many
logical knots and rendering what we have been using understandable.

I will give it a rest for a couple days, and let the testing proceed
unaltered over the week end. Then we will make a tag and move on.

The next part will consist of actually changing the computations, which will
change verification experiments results. We will most likely let go of rigorous
backward compatibility soon after the upcoming tagged version.

Cheers.