formatting changes; latex tables fixed up
So the Oliver magic is to use a command that only affects the latex table formatting.
To make it all look nice I adjusted the (latex) column widths spacings separately by table.
Note I push them a bit into the right margin by having the fractions add to 1.025
The html is unaffected.