Update ctrl_get_gen.F
Originally only the first two time variable tau u and tau v control records were disabled.  We think this is because the long ago someone noticed that the optimization tends to make very large adjustments to the first one or two records for reasons that are not entirely clear.  We recently found very large adjustments in the first two control records of all of the other atmospheric controls.  This 

The DISABLE_FIRST_TWO_ATM_CTRLS_XX flag simply extends the original logic applied to tau u and tau v fields to all the time variable atmospheric controls.