A hack to fix the "-nocat4d" option with new version of TAF
With new version of TAF (currently 4.2.7), *.flowdir files are now
assumed to be free-format source files (since they don't have either .f
or .F suffix); Up to version 4.1.1 they were assumed to be fixed format.

And when using "-nocat4d" option, we send all individual files to TAF
(instead of a single file: ad_input_code.f); so we need to explicitly
add the "-fixed" option to the TAF command to get the *.flowdir processed
correctly.