Third set of adjointable sea ice growth + solve4temp codes
These are from Winter 2011 and were my attempt to merge my sea ice codes into the main branch.
The first attempt failed because of a missing cap on AREA. After that G. Forget decided to
rewrite seaice_growth and from that point on my version of the code lived in contrib.