The sixth version of the adjointable sea ice growth codes
This version of the sea ice growth code was apparently the version that Arash Bigdeli first examined.
It is very similar to the fifth version except that it has the old sea ice plume codes in it
which seem to have been removed at an earlier stage.  I cannot find this particular version anywhere
on my own repositories which suggests that he or An may have added in the sea ice plume codes from version five.