add code to include surface freshwater flux in balancing
This first version uses PmEpR of the previous time step. This is not
optimal because it will break restarts, but requires only minimal code
changes. Properly EmPmR should be used, but then obcs_balance_flow
needs to be called at the end of do_oceanic_phys, when EmPmR is available.