update code to be more consistent with main version
new code contains only differences specific to the experiment, but
also contains an extra call to new subroutine
s/r obcs_balance_and_tides at the end to maintain verification
experiment accuracy. This is a hack and needs to be revisited, but
currently this is the only way not to break this experiment.