Comment out OBCS_PARM02 (#undef ALLOW_ORLANSKI )
This second namelist is only read-in when ALLOW_ORLANSKI is #define.
Generally, with most compilers, this does not pose a problem to read-in
the next namelist, ignoring this un-read namelist; However, with current
version of open64 comiler on engaging (v 4.5.2.1), it fails to skip the
un-read namelist when trying to read the next one.