Test for maskC = 1 (instead of hFacC > 0)
This is a minor modification that might help Autodiff compiler
(since maskC is kept fixed but hFacC can change when using Non-Lin Free-Surf).

Also: use kLowC to test for bottom cell ; and change to first column "C" (upper-case)
to indicate a comment (whereas "c" is used for commented out line of code).