replace endrec by diffrec where it makes sense
this change fixes a bug that makes it impossible to run an AD
simulation from an arbitray pickup file: This routine assumes
startrec = 1 (which is not true when starting from a pickup) and
endrec can become quite large, so that the number of records written
to fnamegenIn (in S/R ctrl_init, calls ctrl_init_ctrlvar with
diffrec->ncvarrecs which then calls ctrl_set_globfld_xy, where file is
written with ncvarrecs = diffrec records) are not enough. diffrec =
endrec-startrec+1 appears to the be correct number of records.