suppress writing meta files in reverse mode in adread_* files
by specifying a negative record number. This assumes that the meta
files of adxx-files (gradients with respect to controls), etc. have
been properly written in the initialisation phase and fixes a bug
where the parameter nrecords in the meta file does not correspond to
the number of records actually in the adxx-file.