fix small issue #419
As discussed in issue #419, the "nf-config" call is, at the end "safe" (since
we check for return value in second part of the test, even when command is
not found, the logic works as expected) but a bit annoying (error message).
By re-directing the error to /dev/null, the annoying message (if any) is gone.

Also update a little bit the comments at the top of this optfile.