Turn off CPP option: ALLOW_AUTODIFF_INIT_OLD
This CPP option was added by @gaelforget in Dec 2015 to enable to reproduce old results.
Although it was not well documented as "fixing a bug" when left #undef (and, when #define,
"adding an old bug", from mid 2004), it is certainly the case.
The corresponding (wrong) code in autodiff_init_varia.F which is added back when CPP
option ALLOW_AUTODIFF_INIT_OLD is defined, does 2 wrong things:
1) it resets 2-D fields "dEtaHdt" after loading this field from pickup-files. This breaks
  the restart and affects all four AD global_ocean_cs32x15 experiments (all starting from
  pickup files that do contains non-zero "dEtaHdt" field).
2) it reset 2-D field "EmPmR" when pkg/exf is compiled. This will not cause any problem
  (and has no effect in any of the four AD global_ocean_cs32x15 experiments) if pkg/exf
  is used or if periodicExternalForcing=.TRUE., (in these 2 cases EmPmR is then loaded later);
  However, in case a constant in time EmPmRFile is used (periodicExternalForcing=F)
  and pkg/exf is compiled but not used (useEXF=F) this will zero out the fresh-water forcing
  for the full simulation (and not only the first time-step).

As a consequence of removing this old bug, update all 4 ADM and all 4 TLM reference output
files from reference platform "villon.mit.edu".