store forcing fields and some ptracer fields and gchemTendency
main change: a new file contains store directives for forcing
fields to be included in the forward_step AFTER load_fields_driver etc
have been called. This avoids an extra call of load_fields_driver in
forward_step_ad (and hence all the stores of the intermediate
variables). So far it does not work when thsice is compiled
minor change: extra store directives for gchemTendency in an effort to
suppress calling do_oceanic_phys in some experiments, incomplete