Much more accurate pLoc for basal melting
For ice-shelf thermodynamics, change the expression of local pressure (in dBar) at the
bottom of the ice-shelf to be simply the weight of the ice-shelf above. This is more
accurate than previous expression based on (truncated) ice-shelf draft, specially when:
 1) shelficeLoadAnomaly is large (static contribution), e.g., because of hFacMin
    truncation on R_shelfIce ;
 2) significant adjustment of sea-level below the ice-shelf occurs (dynamic contribution).
Also store prognostic variable "shelficeMass" (using TAF store directives) to avoid
major recomputations.