Check and stop if OLx,y < 2
Stop if momStepping is used with Overlap size smaller than 2.
This reflects more a practical limitation (some loop range in few momentum S/R)
than an algorithmic limitation (solve_for_pressure RHS) which should allow to
use overlap size of 1 providing momAdvection=F and viscAh=0.
Note that overlap size of 1 can still be used for offline / tracer advection
depending on which advection scheme is used (and this is checked for in gad_check.F).