Switch to #define ALLOW_OBCS_STEVENS
Note: the Stevens OB code is not used in this set-up so that compiling this code is not
necessary (and does increase the number of written tapes) but does not change results.
The ONLY reason it is turned on here is to check that this code compiles in Adjoint set-up.