switch tutorial_global_oce_optim/code_ad to generic controls
This is a bit of a larger operation, because this experiment uses
customized code and a unique control variable HFLUXM (mean heat
flux). The strategy here was to replace this control variable by the
generic xx_qnet with a zero period and hard-wire xx_gentim2d(...,1) in
the customized cost_hflux.F to replace the global variable Qnetm,
which had been introduced only for this experiment. This involves a
few extra changes:
- local genmake and ad_optfile are no longer necessary (no more
  HFLUM_CONTROL), also local copies of some central files are no
  longer necessary.
- cleanup of remaining local files
- pkg/ctrl/ctrl_map_forcing.F requires an addition #ifdef ECCO_CTRL_DEPRECATED
- note that with the generic control variable I needed to set the
  CPP-flag ALLOW_PACKUNPACK_METHOD2, because there is something fishy
  with the default methods and reading the weights file, probably
  because we use MDSREADFIELD_2D_GL there. Needs to be sorted out later.