fix up all routines in ecco_toolbox.F so that they stand a chance
of working. Since some of the routines are not used anywhere, I cannot
be sure about that, but I removed all unconditional stops.

Fix size of localmask (to be nnzsiv4) in cost_gencost_seaicev4.F and
use 2D maskInC instead of maskC for convenience, to be tested. Also
replace "1" by nnzsiv4 where I found it appropriate.