store dirs should not be enclosed by ifdef ALLOW_GENTIM2D_CONTROL
The reason to add ifdef/endif around store directives is to make sure
that the variables to be stored are defined. Here, the variables (e.g.
bdot_streamice) are defined independently of
ALLOW_GENTIM2D_CONTROL (and the ctrl-pkg entirely) and using this
flag would require including CTRL_OPTIONS.h unnecessarily in some
files, so I remove it. Let us hope for the best.