replace USE_OLDSTORE by complentary AUTODIFF_USE_STORE_RESTORE flag
- AUTODIFF_USE_OLDSTORE_2D/3D/SEAICE/EXF/OBCS flags that turn off the
store/restore method are removed.

- Instead a more meaningful flag AUTODIFF_USE_STORE_RESTORE turns on
the store/restore method, undefined by default. For obcs there is a
separate flag AUTODIFF_USE_STORE_RESTORE_OBCS because obcs fields
currently cannot use the WHTAPEIO method so that it still makes sense
to use the store/restore method; however, only the u/v/t/s files are
subject to that and it is likely that we will get rid of this flag and
code as well.

- All verifcation experiments now use the default
method (AUTODIFF_USE_STORE_RESTORE undefined) with WHTAPEIO, except
for tutorial_dic_adjoffline, where results change for unknown reasons
related to storing of 3D fields (AUTODIFF_USE_OLDSTORE_3D code);
tutorial_tracer_adjsens does not use WHTAPEIO, because that changes
the results due to a suppressed called of s/r
convective_adjustement_ini (will be addressed in a separate PR).