put (wrong) key computation within ifdefs and add comment about it
the routine is called from gmredi_calc_tensor three times for each
level k, neither of which is taken into account with this
key. This is only relevant if GMREDI_WITH_STABLE_ADJOINT is
defined, which is the case for ECCOv4, so we add the ifdef
around the key computation.