Simplify/clarify local variable names
Use unique pair of local var for AD & TL executable (instead of a different pair
for TAF and Tapenade).
This simplifies (now "make CLEAN" does remove Tapenade executables) and
clarifies the Makefile (as previously EXE_AD & EXE_FTL were set at the top
of the Tapenade Makefile but not use anywhere in Makefile).