Revert changes for multi-threading if ALLOW_AUTODIFF
TAF get confused about BEGIN_MASTER/END_MASTER and does not recognize that shared
variables in common block do not need to be set by all threads (and safer when they
are not). This leads to major recomputations and significant slow-down.
Revert changes from PR #716 regarding multi-threading when pkg/autodiff is compiled.