remove "*.list" by "_ad_diff.list" to be more specific
This sets additional constraints on naming the file(s) that contain
lists of files to shown to TAF, but it is safer as ".list" is not very
specific and a user might accidentally leave a file with suffix
".list" somewhere in one of the multiple dirs where src code is
found. Currently, all "*.list" files in the MITgcm repository are really
"*_ad_diff.list" files, so this change has not problematic effect.