deal with special case case: implicDiv2DFlow=0
In case implicDiv2DFlow=0, the CG2D Matrix with new units is not right
but there is no real need to solve for cg2d_x as the solution is already
known (when exactConserv=T). In this case, just skip the solver call
(and CALC_DIV_GHAT) and put directly the updated solution into cg2d_x.