fix store directives and recomputations in S/R obsfit_cost
It is not clear to me, why in comparison to s/r cost_profiles, the
store directives are necessary. I think that variable obs_modmask
should be an array (but that does not help either) and I think that
the mix of formal parameters and common block variables in
obsfit_active_file*.F does not help.