adjust namelist for better match between input and input.exf
- still, the input.exf output is different, because the initialisation
of gchemSi is different. In pkg/exf it is initialised to a constant
value and in the default way it is initialised the correctic field at
nIter0. This seems to be necessary, because gchemSi used before
gchem_fields_load is called for the first time. Not sure how to fix this.